requirements.txt
Django==2.2.7
django-markdownx==3.0.1
Markdown==3.1.1
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
